requirements.txt
channels==2.3.0
channels-redis==2.4.1
Django==1.11.17
django-ace==1.0.4
django-cors-headers==2.4.0
django-oauth-toolkit==1.1.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-session-timeout==0.1.0
django-suit-redactor==0.0.4
django-user-agents==0.3.2
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.0
python-debian==0.1.36+ubuntu1.1
python-dotenv==0.21.1
pytz==2018.9
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.23.0
requests-unixsocket==0.2.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0

